+++ spec/nacha/record/adv_batch_control_spec.rb
@@ -80,7 +80,7 @@ RSpec.describe 'Nacha::Record::AdvBatchControl', :nacha_record_type do
 
   describe 'instance generates json' do
     let(:record) { record_class.parse(example_adv_batch_control_record) }
-    let(:formatter) { Nacha::Formatter::JsonFormatter.new([record]) }
+    let(:formatter) { Nacha::Formatter::JsonFormatter.new(Nacha::AchFile.new([record])) }
     let(:abcr_json) { JSON.parse(formatter.format)['records'].first }
 
     it 'is well formed' do