+++ spec/nacha/record/adv_file_control_spec.rb
@@ -25,7 +25,7 @@ RSpec.describe Nacha::Record::AdvFileControl, :nacha_record_type do
 
   describe 'parses a record' do
     let(:fcr) { described_class.parse(example_file_control_record) }
-    let(:formatter) { Nacha::Formatter::JsonFormatter.new([fcr]) }
+    let(:formatter) { Nacha::Formatter::JsonFormatter.new(Nacha::AchFile.new([fcr])) }
     let(:fcr_json) { JSON.parse(formatter.format)['records'].first }
     let(:fcr_hash) do
       {
@@ -103,7 +103,7 @@ RSpec.describe Nacha::Record::AdvFileControl, :nacha_record_type do
 
   describe 'instance generates json' do
     let(:record) { described_class.parse(example_file_control_record) }
-    let(:formatter) { Nacha::Formatter::JsonFormatter.new([record]) }
+    let(:formatter) { Nacha::Formatter::JsonFormatter.new(Nacha::AchFile.new([record])) }
     let(:record_json) { JSON.parse(formatter.format)['records'].first }
 
     it 'is well formed' do