+++ spec/nacha/record/iat_foreign_coorespondent_bank_information_addenda_spec.rb
@@ -32,7 +32,7 @@ RSpec.describe Nacha::Record::IatForeignCoorespondentBankInformationAddenda, :na
 
   describe 'instance generates json' do
     let(:record) { described_class.new }
-    let(:formatter) { Nacha::Formatter::JsonFormatter.new([record]) }
+    let(:formatter) { Nacha::Formatter::JsonFormatter.new(Nacha::AchFile.new([record])) }
     let(:json) { JSON.parse(formatter.format)['records'].first }
 
     it 'is well formed' do