+++ spec/nacha/record/sixth_iat_addenda_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe Nacha::Record::SixthIatAddenda, :nacha_record_type do
 
   describe 'instance generates json' do
     let(:record) { described_class.new }
-    let(:formatter) { Nacha::Formatter::JsonFormatter.new([record]) }
+    let(:formatter) { Nacha::Formatter::JsonFormatter.new(Nacha::AchFile.new([record])) }
     let(:json) { JSON.parse(formatter.format)['records'].first }
 
     it 'is well formed' do